Initial setup for LiteRT (TensorFlow Lite) model tests. (#59)
Progress on https://github.com/iree-org/iree-test-suites/issues/5.

This contains two simple test cases for demonstration purposes, one of
which is currently failing due to a regression:
https://github.com/iree-org/iree/issues/19402.

The test suite follows the same structure as the onnx_models test suite
in this repository. Some cleanup and refactoring will be more evident as
this grows. We could for example share the `compile_mlir_with_iree`
helper function between both test suites.