[torch_models] Change module caching behavior (#121)
Before this patch, compiled modules were being cached by default and the
force recompile feature simply deleted this cached directory. The
problem with this behavior is multiple runners sharing the cache
directory can conflict with each other's modules if they are not caching
modules.

This patch makes them create a temporary working directory inside the
artifact directory (can also change it to a user passed flag in future)
and cache modules for the run during that.

This patch also flips the original "force-recompile" flag, and adding a
"cache-modules" flag, which caches the modules between different runs.