Update CPU flags to follow https://github.com/iree-org/iree setup (#123)
The regression was caused because of the data-tiling flip. It's disabled
by default, so we have to explicitly add the flag.

The revision also refreshes the CPU flags to follow what we've been
using in iree repo; update the golden values for torch models.

The sharktank golden values are back; additionally, the revision
tightens the golden values a bit.

Fixes https://github.com/iree-org/iree-test-suites/issues/116

---------

Signed-off-by: hanhanW <hanhan0912@gmail.com>