Add PyTorch operator tests (#114)
This PR adds pytorch operator tests. This follows a similar approach to
ONNX ops, where the tests are defined at a higher level of abstraction
(in this case the tests are defined at the PyTorch level) and
pre-generated ahead of time. The workflow is as follows:

1. Add a new test definition in `generate_test.py`
2. Regenerate the tests `python generate_test.py`
3. Run the tests with `pytest`.

The MLIR representation is generated AOT of time and saved to a file.
Numpy files are used to store the input parameters and the expected
output. A json file is used to bundle these parameters together along
with rtol and atol which are used to compare the expected output with
the observed output.

---------

Signed-off-by: Erick Ochoa <erick.ochoalopez@amd.com>